Implemented `Canvas.selected_frame(x, y, width, height)` to get a subset of a `Canvas`.
The given `(x, y)` coordinates are used for the top-left corner of the
rectangular frame, with the given `width` and `height`.

Added `Canvas.width()` and `Canvas.height()` which returns the `Canvas`
member fields directly and changed the functionality of
`Canvas.get_width()` and `Canvas.get_height()` to first re-calculate the
bounds, then return the value.